- Extraction de la classe ScenarioLine afin de pouvoir la réutiliser dans d'autres endroits - le constructeur de ScenarioLine est passé en Privé pour montrer qu'on doit passer par le constructeur de la fonction statique avant de passer par lui
