+++ business/aircraft/AircraftFactory.java
@@ -4,6 +4,7 @@ import java.util.UUID;
 
 import business.Coordinates;
 import business.flyable.Flyable;
+import business.scenario.ScenarioLine;
 
 public class AircraftFactory {
     private static AircraftFactory aircraftFactory;
@@ -17,9 +18,12 @@ public class AircraftFactory {
         return aircraftFactory;
     }
 
-    public Flyable newAircraft(String type, String name, Coordinates coordinates) throws Exception {
+    public Flyable newAircraft(ScenarioLine sl) throws Exception {
         // Generate a random unique ID with a length 5.
         long uuid = Math.abs(UUID.randomUUID().getLeastSignificantBits()) % 100;
+        String name = sl.name;
+        Coordinates coordinates = sl.coordinates;
+        String type = sl.type;
 
         switch (type) {
             case "BALOON":
